+++ components/lg-base-form-sheet/lg-base-form-sheet.vue
@@ -1,0 +1,27 @@
+<template>
+	<view>
+		
+	</view>
+</template>
+
+
+<script setup>
+	
+	
+	
+</script>
+
+<script>
+	export default {
+		name:"lg-base-form-sheet",
+		data() {
+			return {
+				
+			};
+		}
+	}
+</script>
+
+<style lang="scss">
+
+</style>